+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## 2025-02-07 v3.4.0
 ### Added
-- updateLoginMenuProps: Method within actionsMenu to handle menu updates
+- setLoginMenuProps: Method within actionsMenu to handle menu updates
 ### Fixed
 - Updating menu item name from Login to Logout after a user is logged in and visa versa.